requirements.txt
# stage 01
# langchain
# openai
# faiss-cpu
# pinecone-client
# langchain-openai


# stage 02
langchain_community
tiktoken 
langchain-openai
langchainhub
chromadb
langchain
bs4
#stafe 06
pandas

# stage 07
sentence-transformers
#faiss-gpu-cu12
faiss-cpu

# lg_stage01
langgraph
#langsmith